requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.42.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
apscheduler>=3.10.4
agentql>=1.9.2
python-telegram-bot==20.7

# Development dependencies
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0 